opsss artifactid back to restheart
from http://maven.apache.org/guides/mini/guide-naming-conventions.html:

artifactId is the name of the jar without version. If you created it then you can choose whatever name you want with lowercase letters and no strange symbols. If it's a third party jar you have to take the name of the jar as it's distributed.