completely replaced logic for Schema Transformation and Check RH-53 RH-68
implementation based on ScriptMetata with javascript logic abandoned because of performance and security reasons. 
performance tests shows that execution time of 50 POST pass from 0.5s to about 1 minute
also javascript cannot be easily sandboxed (for instance to avoid infinite loop)

now we can specify a named singleton in the configuration file. the singleton can be referenced by name in the metadata field